Refactored all endpoint construction logic to an endpoint_utils.py file
Also adding package dependencies on AWSSigV4 libraries
